requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-geomap==1.6
django-admin-multi-select-filter==1.0.0
django-admin-numeric-filter==0.1.9
django-admin-rangefilter==0.9.0
django-cors-headers==3.14.0
django-import-export==3.1.0
django-json-widget
django-prettyjson
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
googlemaps==4.10.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.1
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
psycopg2
gunicorn
boto3
pandas
ortools